requirements.txt
# Core (already installed)
hyperliquid-python-sdk>=0.8.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# New for arbitrage system
aiohttp>=3.9.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
